wip(#1099): extract lower_call/method_override
